move internal/search/run/symbol.go to internal/search/symbols
